got filters working with pagination
