+++ src/components/character/CharacterList.js
@@ -7,13 +7,7 @@ import ApiManager from "../../modules/ApiManager";
 
 const CharacterList = props => {
   const activeUser = JSON.parse(sessionStorage.getItem("credentials"));
-  const [characters, setCharacters] = useState([]);
-
-  /* State related to the filter dropdowns */ 
-  const [characterTypes, setCharacterTypes] = useState([]);
-  const [characterSubTypes, setCharacterSubTypes] = useState([]);
-  const [filter, setFilter] = useState({type: "", subtype: ""})
-
+  const filteredCharacters = props.filteredCharacters;
   /* State related to pagination 
   reference: https://www.npmjs.com/package/react-hooks-paginator
   */
@@ -21,11 +15,6 @@ const CharacterList = props => {
   const [offset, setOffset] = useState(0);
   const [currentPage, setCurrentPage] = useState(1);
   const [currentData, setCurrentData] = useState([]);
-
-  const getCharacters = () => {
-    ApiManager.getCharacterList()
-      .then(setCharacters)
-  }
 
   const getHighConcept = (character) => {
     const aspects = character.characterAspects;
@@ -42,35 +31,6 @@ const CharacterList = props => {
     return highConcept;
   }
 
-  const getCharacterTypeList = () => {
-    return ApiManager.getAll("characterTypes")
-      .then(setCharacterTypes)
-  }
-
-  const getCharacterSubTypeList = () => {
-    return ApiManager.getAll("characterSubTypes")
-      .then(setCharacterSubTypes)
-  }
-
-  const handleFilterFieldChange = (evt, {name, value}) => {
-    const stateToChange = {...filter};
-    stateToChange[name] = value;
-    // If they're a PC, set their subtype for them
-    // (there's no subtype for PCs)
-    // Otherwise, wipe any chosen subtype
-    if (name === "type") {
-      if (value === "1") {
-        stateToChange["subtype"] = "6";
-        // If they're changing the character's type, 
-        // clear any chosen subtype info
-      }
-      else {
-        stateToChange["subtype"] = "";
-      }
-    }
-    setFilter(stateToChange);
-  }
-
   const handleDelete = (id) => {
     confirmAlert({
       title: 'Confirm to submit',
@@ -79,7 +39,7 @@ const CharacterList = props => {
         {
           label: 'Yes',
           onClick: () => ApiManager.delete("characters", id)
-            .then(getCharacters)
+            .then(props.getCharacters)
         },
         {
           label: 'No',
@@ -89,101 +49,31 @@ const CharacterList = props => {
     });
   }
 
-  const renderSubtypes = () => {
-    let subtypes = characterSubTypes
-      .sort((a,b) => a.name.localeCompare(b.name))
-    if (filter.type !== "") {
-      subtypes = subtypes
-        .filter(subtype => subtype.characterTypeId === parseInt(filter.type))
-    }
-    return subtypes.map(type => (
-      {
-        key: `type-${type.id}`,
-        value: `${type.id}`,
-        text: `${type.name}`
-      }
-    ))
-  }
-
-  const renderCards = () => {
-    /* Sorting characters alphabetically
-          https://stackoverflow.com/a/45544166*/
-    let characterList = currentData.sort((a,b) => a.name.localeCompare(b.name))
-    if (filter.type !== "") {
-      characterList = characterList
-        .filter(character => character.characterSubType.characterTypeId === parseInt(filter.type))
-    }
-    if (filter.subtype !== "") {
-      characterList = characterList
-        .filter(character => character.characterSubTypeId === parseInt(filter.subtype))
-    }
-    return characterList.map(character => 
-      <CharacterCard
-        key={character.id}
-        character={character}
-        highConcept={getHighConcept(character)}
-        handleDelete={() => handleDelete(character.id)}
-        activeUser={activeUser}
-      />
-    )
-  }
-
   useEffect(() => {
-    getCharacters();
-    getCharacterTypeList();
-    getCharacterSubTypeList();
-  }, [])
-
-  useEffect(() => {
-    setCurrentData(characters.slice(offset, offset + pageLimit))
-  }, [offset, characters])
+    setCurrentData(filteredCharacters.slice(offset, offset + pageLimit))
+  }, [offset, filteredCharacters])
 
   return (
     <>
       <Container text>
-        <Button
-          type="button"
-          onClick={() => {props.history.push("/characters/new")}}
-        >
-          <Icon className="add user"></Icon>
-        </Button>
-        <Dropdown
-          clearable 
-          selection
-          onChange={handleFilterFieldChange}
-          className="type"
-          name="type"
-          id="type"
-          placeholder="Filter by type"
-          value={filter.type}
-          options={characterTypes.map(type => (
-            {
-              key: `type-${type.id}`,
-              value: `${type.id}`,
-              text: `${type.name}`
-            }
-          ))}
-        />
-        <Dropdown 
-          clearable
-          selection
-          onChange={handleFilterFieldChange}
-          className="subtype"
-          name="subtype"
-          id="subtype"
-          placeholder="Character Subtype"
-          value={filter.subtype}
-          options={renderSubtypes()}
-        />
         <div className="header-container">
           <h1>Characters</h1>
         </div>
         <div>
           <Card.Group itemsPerRow={3}>
-            {renderCards()}
+            {currentData.map(character => 
+                <CharacterCard
+                  key={character.id}
+                  character={character}
+                  highConcept={getHighConcept(character)}
+                  handleDelete={() => handleDelete(character.id)}
+                  activeUser={activeUser}
+                />
+              )
+            }
           </Card.Group>
           <Paginator 
-            totalRecords={characters.length}
+            totalRecords={filteredCharacters.length}
             pageLimit={pageLimit}
             pageNeighbours={2}
             setOffset={setOffset}
